Extract Twitter routes in own controller
