+++ src/services/http/controllers/index.ts
@@ -2,5 +2,6 @@ export * from './Events';
 export * from './Interep';
 export * from './Posts';
 export * from './Tags';
+export * from './Twitter';
 export * from './Users';
 export * from './ZkChat';
