+++ src/services/http/index.ts
@@ -212,7 +212,7 @@ export default class HttpService extends GenericService {
   };
 
   initControllers() {
-    ['users', 'posts', 'tags', 'zkChat', 'events', 'interep'].forEach(controller => {
+    ['users', 'posts', 'tags', 'zkChat', 'events', 'interep', 'twitter'].forEach(controller => {
       this.app.use('/v1', this.get(`${controller}Controller`, 'router'));
     });
   }
@@ -262,198 +262,6 @@ export default class HttpService extends GenericService {
 
         await linkDB.update(data);
         res.send(makeResponse(data));
-      })
-    );
-
-    this.app.get(
-      '/twitter',
-      this.wrapHandler(async (req, res) => {
-        const text = await requestToken();
-        const { oauth_token: token, oauth_token_secret: tokenSecret } = queryString.parse(text);
-
-        // @ts-ignore
-        req.session.tokenSecret = tokenSecret;
-        // @ts-ignore
-        req.session.redirectUrl = req.query.redirectUrl;
-        res.send(makeResponse(`${TW_AUTH_URL}?${queryString.stringify({ oauth_token: token })}`));
-      })
-    );
-
-    this.app.get(
-      '/twitter/callback',
-      this.wrapHandler(async (req, res) => {
-        // @ts-ignore
-        const { oauth_token: token, oauth_verifier: verifier } = req.query;
-        // @ts-ignore
-        const tokenSecret = req.session.tokenSecret;
-        // @ts-ignore
-        delete req.session.tokenSecret;
-
-        const text = await accessToken(token as string, verifier as string, tokenSecret);
-        const {
-          oauth_token: userToken,
-          oauth_token_secret: userTokenSecret,
-          screen_name: userName,
-          user_id: userId,
-        } = queryString.parse(text);
-
-        if (!userToken || !userTokenSecret || !userName || !userId) {
-          throw new Error('invalid oauth');
-        }
-
-        const twitterAuthDB = await this.call('db', 'getTwitterAuth');
-        await twitterAuthDB.updateUserToken({
-          userToken,
-          userTokenSecret,
-          userName,
-          userId,
-        } as {
-          userToken: string;
-          userTokenSecret: string;
-          userName: string;
-          userId: string;
-        });
-
-        // @ts-ignore
-        req.session.twitterToken = jwt.sign({ userToken }, JWT_SECRET);
-
-        // @ts-ignore
-        const redirectUrl = req.session.redirectUrl;
-        // @ts-ignore
-        delete req.session.redirectUrl;
-        res.redirect(redirectUrl);
-      })
-    );
-
-    this.app.get(
-      '/twitter/session',
-      this.wrapHandler(async (req, res) => {
-        // @ts-ignore
-        const { twitterToken } = req.session;
-        const signature = req.header('X-SIGNED-ADDRESS');
-        const twitterAuthDB = await this.call('db', 'getTwitterAuth');
-        const userDB = await this.call('db', 'getUsers');
-
-        let auth;
-
-        if (signature) {
-          const [sig, address] = signature.split('.');
-          const user = await userDB.findOneByName(address);
-
-          if (user?.pubkey) {
-            if (verifySignatureP256(user.pubkey, address, sig)) {
-              const sigAuth = await twitterAuthDB.findUserByAccount(address);
-
-              if (sigAuth) {
-                auth = sigAuth;
-
-                // @ts-ignore
-                req.session.twitterToken = jwt.sign(
-                  {
-                    userToken: auth.user_token,
-                  },
-                  JWT_SECRET
-                );
-              }
-            }
-          }
-        }
-
-        if (!auth) {
-          const jwtData: any = await jwt.verify(twitterToken, JWT_SECRET);
-          auth = await twitterAuthDB.findUserByToken(jwtData?.userToken);
-        }
-
-        const json = await verifyCredential(auth.user_token, auth.user_token_secret);
-
-        const {
-          followers_count,
-          verified,
-          profile_image_url,
-          profile_image_url_https,
-          screen_name,
-        } = json;
-
-        const botometerResult = await getBotometerScore(screen_name);
-
-        const reputation = calculateReputation(OAuthProvider.TWITTER, {
-          followers: followers_count,
-          verifiedProfile: verified,
-          botometerOverallScore: botometerResult?.display_scores?.universal?.overall,
-        });
-
-        res.send(
-          makeResponse({
-            user_id: auth.user_id,
-            user_token: auth.user_token,
-            user_token_secret: auth.user_token_secret,
-            username: auth.username,
-            followers: followers_count,
-            verifiedProfile: verified,
-            profileImageUrl: profile_image_url,
-            profileImageUrlHttps: profile_image_url_https,
-            screenName: screen_name,
-            reputation,
-          })
-        );
-      })
-    );
-
-    this.app.get(
-      '/twitter/check',
-      this.wrapHandler(async (req, res) => {
-        const { username } = req.query;
-        const twitterAuthDB = await this.call('db', 'getTwitterAuth');
-        const user = await twitterAuthDB.findUserByUsername(username as string);
-        res.send(makeResponse(user));
-      })
-    );
-
-    this.app.get(
-      '/twitter/user/:username',
-      this.wrapHandler(async (req, res) => {
-        const { username } = req.params;
-        const user = await getUser(username);
-        res.send(makeResponse(user));
-      })
-    );
-
-    this.app.post(
-      '/twitter/update',
-      jsonParser,
-      this.wrapHandler(async (req, res) => {
-        // @ts-ignore
-        const { twitterToken } = req.session;
-        const { status, in_reply_to_status_id } = req.body;
-        const jwtData: any = await jwt.verify(twitterToken, JWT_SECRET);
-        const twitterAuthDB = await this.call('db', 'getTwitterAuth');
-        const auth = await twitterAuthDB.findUserByToken(jwtData?.userToken);
-        const json = await updateStatus(
-          status,
-          in_reply_to_status_id,
-          auth.user_token,
-          auth.user_token_secret
-        );
-        res.send(makeResponse(`https://twitter.com/${auth.username}/status/${json.id_str}`));
-      })
-    );
-
-    this.app.get(
-      '/twitter/status',
-      this.wrapHandler(async (req, res) => {
-        // @ts-ignore
-        const { id } = req.query;
-        const status = await showStatus(id as string);
-        res.send(makeResponse(status));
-      })
-    );
-
-    this.app.get(
-      '/oauth/reset',
-      this.wrapHandler(async (req, res) => {
-        // @ts-ignore
-        if (req.session.twitterToken) delete req.session.twitterToken;
-        res.send(makeResponse('ok'));
       })
     );
